category setup

+++ Backend/src/route.Controller/user.Controller.js
@@ -98,138 +98,149 @@ const create = (req, res) => {
         })
         return false
     }
-    db.query("SELECT COUNT(Email) FROM user WHERE Email=?", [Email], (err, row) => {
-        if (err) {
-            res.json({
-                error: true,
-                message: err
-            })
-        } else {
-            if (row.length > 0) {
+    if (checkPermission(req, 1)) {
+        db.query("SELECT COUNT(Email) FROM user WHERE Email=?", [Email], (err, row) => {
+            if (err) {
                 res.json({
                     error: true,
-                    message: "USer already exist"
+                    message: err
                 })
             } else {
+                if (row > 0) {
+                    res.json({
+                        error: true,
+                        message: "USer already exist"
+                    })
+                } else {
 
-                Password = bcrypt.hashSync(Password, 10)
-                var sql = "INSERT INTO user (Name,Gender,Email,Password,Role) VALUE(?,?,?,?,?)"
-                var parameter = [Name, Gender, Email, Password, Role]
-                db.query(sql, parameter, (err1, row) => {
-                    if (err1) {
-                        res.json({
-                            error: true,
-                            message: err1
-                        })
-                    }
-                    else {
-                        if (row.affectedRows > 0) {
-                            res.json({
-                                message: "Insert success "
-                            })
-                        } else {
-
+                    Password = bcrypt.hashSync(Password, 10)
+                    var sql = "INSERT INTO user (Name,Gender,Email,Password,Role) VALUE(?,?,?,?,?)"
+                    var parameter = [Name, Gender, Email, Password, Role]
+                    db.query(sql, parameter, (err1, row) => {
+                        if (err1) {
                             res.json({
                                 error: true,
-                                message: "Please check "
+                                message: err1
                             })
                         }
+                        else {
+                            if (row.affectedRows > 0) {
+                                res.json({
+                                    message: "Insert success "
+                                })
+                            } else {
+
+                                res.json({
+                                    error: true,
+                                    message: "Please check "
+                                })
+                            }
 
 
-                    }
-                })
+                        }
+                    })
+                }
             }
-        }
-    })
-
+        })
+    } else {
+        res.json({
+            error: false,
+            message: "You don't has permission access this method!",
+        });
+    }
 }
 
 const update = (req, res) => {
+    if (checkPermission(req, 1)) {
+        var {
+            User_Id,
+            Name,
+            Gender,
+            Email,
+            Password,
+            Role,
+            Status
+        } = req.body
+        let message = {}
+        if (isEmpty(User_Id)) {
+            message.User_Id = "id is require "
+        }
+        if (isEmpty(Name)) {
+            message.Name = "please fil in Name "
+        }
+        if (isEmpty(Gender)) {
+            message.Gender = "please fil in Gender "
+        }
 
-    var {
-        User_Id,
-        Name,
-        Gender,
-        Email,
-        Password,
-        Role,
-        Status
-    } = req.body
-    let message = {}
-    if (isEmpty(User_Id)) {
-        message.User_Id = "id is require "
-    }
-    if (isEmpty(Name)) {
-        message.Name = "please fil in Name "
-    }
-    if (isEmpty(Gender)) {
-        message.Gender = "please fil in Gender "
-    }
+        if (isEmpty(Email)) {
+            message.Email = "please fil in Email "
+        }
 
-    if (isEmpty(Email)) {
-        message.Email = "please fil in Email "
-    }
+        if (isEmpty(Password)) {
+            message.Password = "please fil in Password "
+        }
 
-    if (isEmpty(Password)) {
-        message.Password = "please fil in Password "
-    }
+        if (isEmpty(Role)) {
+            message.Role = "please fil in Role "
+        }
+        if (isEmpty(Status)) {
+            message.Status = "please fil in Status "
+        }
 
-    if (isEmpty(Role)) {
-        message.Role = "please fil in Role "
-    }
-    if (isEmpty(Status)) {
-        message.Status = "please fil in Status "
-    }
-
-    if (Object.keys(message).length > 0) {
-        res.json({
-            error: true,
-            message: message
-        })
-        return false
-    }
-    var sql = "UPDATE user SET Name=?,Gender=?,Email=?,Role=?,Status=? WHERE User_Id=?"
-    var parameter = [Name, Gender, Email, Password, Role, Status, User_Id]
-    db.query("SELECT COUNT(Email) FROM user WHERE Email=?", [Email], (err, row) => {
-        if (err) {
+        if (Object.keys(message).length > 0) {
             res.json({
                 error: true,
-                message: err
+                message: message
             })
-        } else {
-            if (row.length > 0) {
+            return false
+        }
+        var sql = "UPDATE user SET Name=?,Gender=?,Email=?,Role=?,Status=? WHERE User_Id=?"
+        var parameter = [Name, Gender, Email, Password, Role, Status, User_Id]
+        db.query("SELECT COUNT(Email) FROM user WHERE Email=?", [Email], (err, row) => {
+            if (err) {
                 res.json({
                     error: true,
-                    message: "Dapicate Email"
+                    message: err
                 })
             } else {
-                db.query(sql, parameter, (err1, row) => {
-                    if (err1) {
-                        res.json({
-                            error: true,
-                            message: err1
-                        })
-                    }
-                    else {
-                        if (row.affectedRows > 0) {
-                            res.json({
-                                message: "Update success "
-                            })
-                        } else {
-
+                if (row.length > 0) {
+                    res.json({
+                        error: true,
+                        message: "Dapicate Email"
+                    })
+                } else {
+                    db.query(sql, parameter, (err1, row) => {
+                        if (err1) {
                             res.json({
                                 error: true,
-                                message: "Please check "
+                                message: err1
                             })
                         }
+                        else {
+                            if (row.affectedRows > 0) {
+                                res.json({
+                                    message: "Update success "
+                                })
+                            } else {
+
+                                res.json({
+                                    error: true,
+                                    message: "Please check "
+                                })
+                            }
 
 
-                    }
-                })
+                        }
+                    })
+                }
             }
-        }
-    })
-
+        })
+    } else {
+        res.json({
+            error: false,
+            message: "You don't has permission access this method!",
+        });
+    }
 }
 
 
@@ -246,30 +257,36 @@ const Delete = (req, res) => {
         })
         return false
     }
-    db.query("DELETE FROM user WHERE User_Id=?", [id], (err, row) => {
-        if (err) {
-            res.json({
-                error: true,
-                message: err
-            })
-        }
-        else {
-            if (row.affectedRows > 0) {
-                res.json({
-                    message: "Delete success "
-                })
-            } else {
-
+    if (checkPermission(req, 1)) {
+        db.query("DELETE FROM user WHERE User_Id=?", [id], (err, row) => {
+            if (err) {
                 res.json({
                     error: true,
-                    message: "id not found"
+                    message: err
                 })
             }
+            else {
+                if (row.affectedRows > 0) {
+                    res.json({
+                        message: "Delete success "
+                    })
+                } else {
+
+                    res.json({
+                        error: true,
+                        message: "id not found"
+                    })
+                }
 
 
-        }
-    })
-
+            }
+        })
+    } else {
+        res.json({
+            error: false,
+            message: "You don't has permission access this method!",
+        });
+    }
 }
 
 module.exports = {
